Add KISS interface
Add a preliminary version of a KISS interface. Note that this interface
does _not_ apply CRC32 checksums on the KISS level, and is therefore
incompatible with previous versions of the interface. The CRC32 check
should instead be performed at the CSP level.

We will probably make CRC32 a default option in a future release.

The underlying USART should call csp_kiss_rx with new data, and should
support a callback function for data not accepted by the KISS interface.