Add rudimentary OSX support
Mac OS X does not support the realtime extensions for posix. Therefore, we cannot reuse the existing posix arch. Insted the pthread queue implementation is used to create a simple locking mechanism with timeouts (pthread_cond_timedwait is the only sync primitive available with a timeout).

There are still problems with resolving weak symbols, but this is a first step.