Updated CHANGELOG, ready for 1.1 release.
Since LibCSP has not received many changes since Feb2012, this
is a "defacto" stable release. Probably the most tested version
to be released so far as its running on AAUSAT3 and GOM-X1 at
the moment.

GomSpace will lock to version 1.1 for flight software, so the
HEAD is once again free to accept new patches.