Fix important bug in csp_buffer_free
The function csp_buffer_free used a critical region to ensure no interrupts occurs while the function csp_buffer_free_isr was called. Logically this is correct, but in some implementations (FreeRTOS + AVR8), this is illegal. The reason is that the CSP critical region maps directly to FreeRTOS portENTER_CRITICAL, and buffer_free_isr maps to xQueueSendFromIsr. This combination does not work on AVR8 and can potentially leave interrupts disabled while in a Task context.

As a consequence, it's now extra important that ALL CSP interfaces ISR routines will use the csp_buffer_free_isr function directly and never the csp_buffer_free.  This is fixed for the csp_if_i2c for now.