Add a new and quite dangerous way to save/load the routing table
The new route_table_save/load functions are designed to be used with
direct flash memory storage on ARM and AVR processors. There is no in/out
validation on the data, so you better be carefull and know what you are doing :)

The use case is:
Subsystems can now ship with standardized firmware and be configured after
boot-up. The user configures the routing table using the CMP ROUTE_SET
interface, and saves a copy of the routing table to flash memory. Saving to flash
memory is per definition a dangerous operation, and should only be performed
in the lab. Not in flight.

Other changes:
	Move loopback route initialization to route_table_init instead of csp_init
	Remove unused route table lock
	Declared route table static
	Allow route_table_init to be called multiple times