csp: drop support for Windows and OS X
The Windows and OS X archs have never been primary citizens of CSP. We do
not use them at GomSpace and do not have the infrastructure to do proper
development and release testing of the archs. So I prefer to remove them
entirely, rather than having operating systems with unknown status in the
tree.