can: socketcan: enable blocking writes for socketcan
By default, SocketCAN interfaces have a txqueue length of 10 frames. If
no buffers are available, writes to the interface returns ENOBUFS
immediately, without blocking. See this thread and others on the
socketcan lists for some background:

http://socket-can.996257.n3.nabble.com/Solving-ENOBUFS-returned-by-write-td2886.html

Our current SocketCAN interface detects the ENOBUFS error and retries
the transmission with a 10 ms delay. This is quite inefficient because
the queue often runs empty while we wait, and as a result our CAN speeds
are much lower than wire speed.

This patch lowers the CAN socket send buffer to the minimum allowed
value (2048), to ensure that the transmissions are blocked at socket
level and not dropped by the interface.

My testing shows a transfer speed increase of 2.5x - 5x in FTP transfers.

Signed-off-by: Jeppe Ledet-Pedersen <jlp@gomspace.com>