rdp: Open connections before receiving first packet
Previously we had to initiate a new RDP connection with a request containing a non-zero length packet. However, for streams this should not be necessary. So moving the conn-queue push to earlier will allow for a service to start transmitting as soon as somebody connects to the port.

Also i have removed the call to srand() since i feel this is something that is best done outside of the RDP stack. Furthermore it resulted in using the same sequence number if two connections were made within the same millisecond