router: added connectionless callbacks
Rationale: For simple request/reply stuff, like echo and friends, an
entire server task with statefull connections should not be needed. We
have already added support for connectionless sockets, but a task is
still required in order to accept a packet from a socket.

This feature changes that by registering a callback function on the
socket using csp_socket_set_callback(). This will set the
CSO_SO_CONNLESS_CALLBACK flag which causes the router to call the
function upon accepting a packet.

The benefit is reduced memory usage by avoiding an extra task. It is
also simpler to setup CSP, since the user does not
need to handle the csp_service_handler(). On systems with a lot of
simple traffic (such as param get/set) this will also benefit
performance and response times since the conn system is avoided together
with the context switch.

The disadvantage is that the router task can call external functions,
which could cause problems if the function calls a blocking call for
too long. However for simple request/replies this should be fine. Since
this feature is considered advanced, we must rely on users to understand
the implications of using callbacks.