rdp: Decrease number of reserved slots in RX queue
When deciding wether to ack a packet, we previously would consider the
RX queue as available if there was enough room for window_size * 2
packets.

However with reduced conn_queue_length (the default setting is now 10 in
our setup), this would reduce the maximum allowed window size to 4.

Considering the fact that it is wastefull to reserve this much buffer
space, we have reduced the number to only a single time the window_size.
This would allow the window size to approach the conn_queue_length.

The user should be mindfull that using a window_size of 10, with a
queue_length of 10 could result in dropped packets, since some packets
could be already in the air. So setting window size a bit below is
always recommended.

This is partially in response to issue #109