dedup: Decrease timeout to 100 ms instead of 1s
The reason for this is that when transferring data over redundant
deduplicated radio links at high speed with RDP enabled, we acutally
want retransmissions of duplicated frames faster than one second.

Given the fact that CSP dedup was actually meant to removed duplicate RX
packets when multiple receivers are running in parallel, 100 ms is still
plenty of time to syncronize the received messages at the receiver (and
remove the duplicate).