Enable custom server IP ports
Previously, the IP ports where the IP interface sends/receives packets
were hardcoded to be 60314. This is an issue when testing locally,
as two clients cannot listen on the same port on a local machine.

Thus, there was a slight API change where the csp_ip_config_t
structure now has a TX port, which is the destination port for outgoing
packets.