Add IP interface example
Configure with --enable-examples and --enable-if-ip to use.

The example simply sends packes from a client task to a server task.
Since they should be running different instances, the example
was designed to be executed with a single argument, either 1 for
client, or 2 for server (representing their CSP addr).

Further, there was a small internal change to how logging is done, as
we don't want stuff printing in network byte order.