Update IP example for remote access
In order to support running the IP example over a network, an
extra REMOTE_ADDR argument was added so that you may directly
specify the IP address of the other remote machine. This will allow
users to run the IP example on two different machines on a network.
This has been tested across macOS and Ubuntu.