udp: Support multiple instances of UDP interface
With this we now take a csp_if_udp_conf_t struct as argument, where the host, rport and lport must be set. This struct is also used as static allocation for the rx thread and the peer addr variable. So the interface avoids dynamic allocation. The interface name should be changed after the interface has been intialised, so it works with routing tables