init: Make dedup a runtime configuration
The deduplication feature is now a runtime setting which can be set to
  CSP_DEDUP_OFF,
  CSP_DEDUP_FWD,
  CSP_DEDUP_INCOMING,
  CSP_DEDUP_ALL

Also, the defualt value of CSP_VERSION has been removed from the meson_options and are now hardcoded as 2. So if you want csp 1, you can now only do it by modifying your call to csp_init.