wscript: Add --version2 build option
We need `CSP_VERSION` defined in order to build and make the code
operate as either CSP version 1 or 2.  meson build already supports
this with the `version` project option with default to 1.

For waf, we add --version2 build option so that we can default to
version 1 without it, or specify it when building version 2.

This fixes #206.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>