src/csp_id: Convert to unsigned int
The commit f84704641a75c109b8 replaced the defines to the functions
but the value returned by the functions are declared as signed int
even though the values never be negatives.

This fixes the following error.

> ../src/rtable/csp_rtable.c: In function ‘csp_rtable_parse’:
> ../src/rtable/csp_rtable.c:63:16: error: comparison of integer
> expressions of different signedness: ‘unsigned int’ and ‘int’
> [-Werror=sign-compare]
>    63 |   if ((address > csp_id_get_max_nodeid()) || (netmask > CSP_RTABLE_MAX_BITS) || (via > UINT8_MAX) || (ifc == NULL))  {
>       |                ^
> cc1: all warnings being treated as errors

This only happens with waf build system because meson build system
doesn't set -Wextra or -Werror.  The following commit will sync the
compiler options used by the build systems.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>