include/csp/csp_util: Add IS_ENABLED macro
IS_ENABLE(CSP_USE_RDP), for example, expands to 1 if defined or 0
otherwise even if not defined.  This allows us to use if-else
condition statements in C functions rather than using #ifdef. It makes
the code much cleaner to read, makes compilers verify the code in both
the true and the false clauses.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>