meson: Make compiled-in log level selectable
This commit make compiled-in log levels selectable with meson
configure.  The default log level is set to the same 'debug' level as
waf build system.  You can use

    meson configure -Dlog-level=info

to select which level you want to compile in.  The levels are

    none
    error
    warn
    info
    debug

All levels above the level you selected will be compiled in.  e.g) If
you select 'info', the above three levels, 'error', 'warn', and
'info', are compiled in.  Select 'none' if you don't need any.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>