static mem: Move back to a compile time qfifo_len definition
While moving towards fully static allocation of everything inside CSP. The choise has been made to skip the runtime initialization in order to also simplify setup of the library.

If csp is to ever be used as a dynamically loaded library, it could simply be built for small, medium and large systems by changing compile time options. Similar to other multilib distributions.

The advantages of more compile time configuration outweighs the slight disadvantage to binary distribution of this library.

Several other areas will be moved to the same methodology.