meson.build: Downgrade to 0.45.1
We currently support Ubuntu 16.04, 18.04, and 20.04.  Unfortunately,
Ubuntu 16.04 has Meson 0.29, which is too old to support.

We decided to support Ubuntu 18.04, which has Meson 0.45.1.  Thus,
downgrade to that version.

In order to downgrade and ensure that we only use the features
available in 0.45, we give a version number in project().  This
revealed that we have two features to remove:

- dict
- install kwarg from configure_file()
- meson compile sub-command

We use dict for log level selection.  The dict is now converted to a
simple but verbose if-elif-else conditions.

install kwarg is removed from configure_file() but because we use
install_dir kwarg, the default behavior is the same as "install:
true".

compile sub-command was added to Meson 0.54.  So this has to go as
well.  We can simply use ninja instead.

No functionality should have been changed.

This fixes #238.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>