buildall: waf: Remove python binding
Python binding doesn't build right now.  This is because csp_init()
has been changed in the commit a24701e8c11483.  The fix should be easy
and similar to the commit cfd20346527ea, but I don't really use it and
don't have test code to prove my change is OK.

So, for now, just remove it from buildall.py.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>