static: Add csp_thread_create_static() for POSIX
This is a RFC.

I came up with this implementation of csp_thread_create_static() for
POSIX in the course of Zephyr port.

**Empty struct for unused type**:

I came up with `empty{}` for unused type.  Is it acceptable?

**Use of abort()**:

As we discussed we don't want to fail from static version of
functions, thus I just put it abort() in it.  Do you think it's
acceptable?

**Order of Function arguments**:

For argument order, I think this version is cleaner but we should
follow the non static version for compatibility.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>