zephyr: Add symlink zephyr pointing to contrib/zephyr/
Zephyr's build system currently search hard coded `zephyr/module.yml`
for each module and integrate to the build system when found.  Thus,
in order to make libcsp a Zephyr module, we have to have zephyr/
directory at the top level directory.

But having platform specific directories in the top directory is too
cluttering.  Thus we've decided to have platform specific directories
under contrib/, and as a special case for Zephyr module, a symbolic
link is placed at the top.

Put libcs module entry in your west.yml:

    - name: libcsp
      url: https://github.com/libcsp/libcsp
      revision: develop
      path: modules/lib/libcsp

With this in place, you can build Zephyr with libcsp like:

    west update
    west build -b qemu_cortex_m3 modules/lib/libcsp/zephyr/samples/arch/

This commit is made separately so that we can easily revert when we
don't need it.

This is discussed in #222.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>