doc: Remove leading and trailing whitespaces
This is a part of rst to gfm conversion.

Leading whitespaces are misleading pandoc to make them blocks. So this
commit removes them.

For the code blocks, tabs don't seems to work. Replace with spaces.

We also have some trailing whitespaces we don't need.  This commit
also remove them as well.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>