meson.build: Use Meson's Python module for building extensions
Meson 0.46 or later provides a python module.  This module allow us to
build a python extension with extension_module().  Replace our own
target with shared_module() with extension_module().

This target generates the final binary with the appropriate naming
convention.  Because the module used to be named "libcsp_py3" this
commit keeps the exact name.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>