cmake: Build object libraries with -fPIC when BUILD_SHARED_LIBS
These object libraries are linked to the final library.  When
BUILD_SHARED_LIBS is ON, we must build them with PIC.  Otherwise
compiler error out at link time:

    /usr/bin/ld: liblibcsp.a(csp_debug.c.o): warning: relocation
    against `csp_debug_hook_func' in read-only section `.text'
    /usr/bin/ld: liblibcsp.a(csp_buffer.c.o): relocation R_X86_64_PC32
    against symbol `csp_debug_level_enabled' can not be used when
    making a shared object; recompile with -fPIC

To build PIC objects, we set a property POSITION_INDEPENDENT_CODE for
the targets.

ref: https://gitlab.kitware.com/cmake/cmake/-/issues/17356

When we build the static library, all depnedencies propagates to
examples/zmqproxy because static library can't be linked against them.
But now with the shared library, dependecies private to libcsp doesn't
propagates to the linking exectutables.  This commits add pthread and
libzmq for zmqproxy dependencies.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>