cmake: Fix building shared lib on Linux
This always confuses me.  GNU Autotools and Meson call host machine to
mean the machine to run the final binary.  CMake, OTOH, calls a build
machine as a host system.

We want to build the shared library for the python binding, which is
supported by Linux, Windows, and macOS, I assume.  But, I don't have
other OSes nor we run them on CI.  So, I'm only adding Linux for this.

Rename CMAKE_HOST_SYSTEM_NAME to CMAKE_SYSTEM_NAME so that we build
the python binding when we target for Linux.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>