windows: Add examples/csp_server_client_windows.c
csp_server_client requires arch specific thread creation functions.
With xamples/csp_server_client_windows.c, csp_server_client can now be
compiled for Windows.

I use uintptr_t for the return variable because it what
_beginthreadex() returns. We don't have anything to pass to the
function except the thread entry function.

The wrapper was using 0 for the last parameter but it's actually a
pointer to an unsigned int. So I've changed it to NULL. That's what
the doc says:

https://docs.microsoft.com/en-us/cpp/c-runtime-library/reference/beginthread-beginthreadex

    thrdaddr

    Points to a 32-bit variable that receives the thread
    identifier. If it's NULL, it's not used.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>