csp_thread: Remove csp_bridge_start
csp_bridge_start() is a function to start a bridge task with
csp_bridge(), which is just a while loop calling csp_bridge_work().
In order to remove csp_thread from libcsp, we remove this API.

Because no sample code is using csp_bridge_start(), we just simply
drop it.

This closes #289.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>