Got rid of global csp_conf.address
The global address was in the way of having multiple addresses, which is necessary for responding to broadcast traffic on each subnet. So it had to go.

Along with it it also deprecates the use of csp_get_address() which is used in userland a few places already.

We need to work towards the notion that a node can have multiple addresses.