Added yaml configuration format
For our Posix/C based tools (satctl), we have a plethora of command line options to try and configure various interfaces. It doesn't get better now that each interface should have its own address and netmask. So in order to ease configuration i have added a small yaml file.

I wanted to add this to CSP because i belive other tools may benefit from this too.

It depends on libyaml, and automatically byilds the functions if it's available on the system. It also depends on malloc.

oh and btw. the implementation is lossy, i haven't bothered cleaning up string's and such, because it's an init time tool only and mostly running on PC. I should probably add a note to the source file about this, before someone comes yelling at me :)