Build libyaml as meson subproject
If the development headers of libyaml is not installed on the os, use libyaml from the local subprojects.

At the time of writing there is a bug in the latest ubuntu packages, preventing libyaml from installing correctly, so adding it directly as a subproject is a simple fix.

libyaml is cmake, but meson can import it anyways. This may not last forever, since most systems should be able to have libyaml.