Fix missing weak functions with meson
For the internal weak functions to be available to the linker, the archive needs to be linked with the -Wl,--link-whole flag. In meson this is done either by setting link_whole : csp_lib in the declare_dependency. But this has one disadvantage if multiple other libraries depend on csp_dep, because it would be included in each library and cause a multiple definition when the final program is linked.

As a solution to this, meson added dependency().as_link_whole() in version 0.56, which is supposed to be called only from the outer level executable()