compat: Remove shared lib support for builtin yaml
Building the bundled libyaml with shared library support uses a feature in meson where it is possible to pass options to a cmake subproject. This feature is too new for the current version of meson on github. So it's commented out for now, and will be enabled again when github gets newer meson.