Remove tasklist, and some of csp_system
The tasklist was only supported on freertos, and it is an old and deprecated interface for debugging with freertos. It pulls in a lot of string handling functions and should not be required.

I have kept csp_memfree_hook, because it's a very simple call and it's supported by most OS'