CSP_DEBUG -> CSP_HAVE_STDIO
Now it's possible to entirely disable stdio (still todo is a few scanf in rtable_parse)

CSP_DEBUG was always set to 1, and have been removed in favor of the new name. If CSP_HAVE_STDIO is set, it uses the default csp_print_func to perform the prints, which can be overridden by the user.