meson.build: Move libc check before source file search
Check all global dependencies before descending into source directory.
This allows you to use the dependencies already found in the subdirs.

We currently don't use partial compilation with subset of include
paths, so this change is actually noop.  This commit prepare for it.
And it's good to group the checks at one place.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>