conn_find_existing: Remove input filter on dst
We only call conn_find_existing, after we clarify that the packet `is_to_me` in the router task. Incoming connections an arrive on any address, so we dont care. The incoming connection is only still unique based on the source id and the source port. So we need to add this again, actually