conn: incoming filter fix
We now filter differently based on who started the connection. For outgoing connections, the source port is unique, which enables outgoing packets to broadcast addresses.

However for incoming connections the source address and port is singular, and unique.

Note that for outgoing connections, the source and destination address is now unnecessary. Which means a routing lookup in the csp_connect() call can be removed.