csp_iflist: Remove stdio.h
csp_bytesize() used to use snprintf() to convert an int to a string.
We should pass numbers as-is to csp_print() so that the conversion
should be done by csp_print, not libcsp proper.

This allows us to remove yet another stdio.h dependency.

This commit also removes the double usage from csp_bytesize().  This
is the only place we use double in libcsp.  We should avoid floats and
doubles in libcsp to support MCUs without a FPU.

This commit is an API breaking change as well as an UI change because
this change eliminates the number after the decimal point.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>