split horizon: Add check on all outgoing packets
With this we now prevent packets being sent back to, not just the same interface, but the same segment it was received from.

This is useful to prevent routing loops when multiple interfaces sits on the same network segment. (Link layer redundancy)

This fixes #336

Note: We still need to do someful useful with the routing table because this currently does not have this feature.