Cleaup memory leaks from csp_yaml
At first i didn't think this was a problem. Because yaml configuration is only used on heavy machines where we could "live" with the leaks of a few strings for some init code.

Hoever, when debugging other memory leaks with valgrind, it's annoying to look at yaml leaking too. So i went ahead and fixed these too.