conn: atomic_compare_exchange_strong
The _weak function can cause sporadic failures to compare correctly, and will leave a no-op. And requires a retry on the same operation in order to funciton. In csp_conn_allocate, we are not retrying, but skipping to next connection candidate. Which would cause it to erronously skip all connections, when they are in fact free.

A change to _strong changes the behaviour, and is required to always return the correct comparison.