Create new port of libcsp to our hardware
Previously, libcsp was changed to the point where it was unusable on any system except Athena.
This new commit takes all of the old drivers for athena, namely CAN and USART and places them such that this version of libcsp will still compile and run on linux.
The reason for this is to support the same libcsp master branch on both the groundstation and the satellite, allowing us to use generic interface features on both