Use explicit csp_packet_t * return
Every call to csp_buffer_get is treated as a csp_packet_t. Also, csp_skbf_t used the size of csp_packet_t to declare the size of its array, effectively treating it as a csp_packet_t.
This change makes the convention explicit. Now getting buffers is exclusively for getting packets (suggest a rename?) but it preserves the generic internal queue API.