nomtu: Remove the usage of the interface MTU field
After leading to much confusion over the years, the mtu field has now
been removed. This is made possible with the recent change to the
packet structure having a compile time defined size. This means we dont
need a runtime configuration field to check against, we can simply check
against sizeof(packet->data) now.

All RX functions need to check for overflow of the packet data field.
All TX functions that is askd to transmit a csp packet larger than their
underlaying layer cannot handle, should drop the packet.

In-the field MTU analysis is done by sending larger and larger ping
packets over the network to determine what the end-to-end MTU is.

Interfaces that support "infinite mtu" are:
    CAN CFP 2.0 (begin and end flag)
    KISS (begin and end flag)
    I2C (begin and end flag)

Interfaces that are limited:
    CAN CFP 1.0 (uses 256 frame remain field, so 2^255 * 8 = 2048 B)
    ZMQ (supports 2^29 bytes at least)
    UDP (limited to 2^16)