csp_buffer_get_always: A new method
From now on, csp_buffer_get will reserve two (2) buffers for critical incoming
messages. These remaining buffers can be used only by calling
csp_buffer_get_always() - This should be reserved only for such places where
not getting a buffer would be detrimental to the reliability of the system.
An example is the CAN input buffers. A failure to retrieve a buffer here means
the system gets non responsive. We handle this type of error by calling
csp_panic() a new hook thats just added. This function must not return,
and should reboot the system or the program running CSP to recover
responsiveness of the system.