Remove weak dummy implementation of csp clock:
Sometimes thease weak implementations end up being selected by the
linker based on the order of linking. This makes the weak attribute more
of an annoyance than a help.

It is easier to explain to users that they will have to implement dummy
hooks in order to link than to debug a runtime issue when suddenly
the weak hooks have been selected over the user supplied hooks