pycsp: Heap allocate socket, close after use
Previous implementation used a pointer to a stack-allocated socket
that was instantly destroyed, leading to a segmentation fault when
used. Changed to use the Python allocator instead.

It also attempted to destroy the socket using the csp_close function,
which is only for connections. Changed to use an appropriate close
function for sockets.