pycsp: Remove address from init, add other fields
CSP 2.0 removed address from the global csp_conf and so it does no
longer make sense to accept it as an argument to pycsp_init. Also,
pycsp_init only allowed to specify some of the fields in csp_conf,
which has now been changed to accept all fields in the csp_conf_t
type.