cmake: Fix python binding option, change to py3
The previous option name for the python bindings,
'enable-python3-bindings' is not a valid varaible name in CMAKE.
Also only Python3 is supported forwards so the CMake package Python3
is now used.

Due to the 'WITH_SOABI' option being specified, the MODULE option
was also explicitly specified as 'WITH_SOABI' can only be used with
'MODULE'.