pycsp: Use correct type specifier for arguments
Ensures that the correct type specifier for argument parsing is always
used, both with regards to size and signedness. This mainly changes
the use of 'h' (short int) to 'H' (unsigned short int) and 'i' (int)
to 'I' (unsigned int). It also replaces uses with 'n' (Py_ssize_t) with
'k' (unsigned long)