examples: Add default intarface setting to all interfaces
CSP now supports having multiple default interfaces due to the
following commit.

   commit 947d810
   subject Support multiple default interfaces

Therefore, a flag called 'is_default' has been added to each
interfaces.

However, in the current csp_server_client example, this flag is not
set. Therefore, if RTABLE is not used, there is no default interface,
and the packet will not be sent anywhere.

Users can decide whether to set each interface the default, but in
this commit, it will change all interfaces to the default.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>