examples: Fix wrong server address settings when not use RTABLE
In the current csp_server_client example, if we don't use the RTABLE,
we cannot test it correctly using the default interface. This is
because without RTABLE, server address and address are always set to
the same value. In this commit, the server address is set to the
given client address if and only if the default interface does not
exists, meaning a loopback test, reagardless of CSP_HAVE_RTABLE.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>