Example: Create server, client, and loopback modes
It's unclear when and how parameters `-a` and `-r` should be used.
The logic of how the client and server threads are started prevents
modifying the client and server's addresses.

Refactor the parameters and logic so that there are well defined
server, client, and loopback modes. A server just needs a server
addresses specified so that it knows how to listen for messages; a
client needs a client address so that it knows how to listen and a
server address so that it knows where to connect to; when neither
client or server addresses are specified both server and client run
on the loopback interface.

This change to the parameters should make it clearer to the user how
to run the client and server in different configurations, and the
logic for starting client and server threads matches the parameter
usage.