route: Accept packets to all the nodes addresses
This allows pinging a node on all its assigned addresses regardless of
which interface it arrived on. Previously CSP would only accept traffic
to the address associated with the interface of the incoming packet.

sendto_reply: If the packet is received on the L2 broadcast address,
reply with the physical interface of the response.