cmake: Add option to enable reproducible builds
Introduce the CSP_REPRODUCIBLE_BUILDS option in CMake. When
CSP_REPRODUCIBLE_BUILDS is set to 1, reproducibility for both static
and shared libraries will be enabled.

Multiple invocations of the following command should generate the same
BuildID and MD5 Hash for the binaries:

    $ cmake -S . -B builddir -DCSP_REPRODUCIBLE_BUILDS=1 && cmake --build builddir
    $ file build/libcsp.so
    $ md5sum build/libcsp.so

Note: By default, CMake embeds the build directory in the binaries,
resulting in non-deterministic builds. To address this, use
`CSP_REPRODUCIBLE_BUILDS=ON` or `CMAKE_SKIP_RPATH=ON` as follows:

    $ cmake -S . -B builddir -DCSP_REPRODUCIBLE_BUILDS=ON -DCMAKE_BUILD_RPATH_USE_ORIGIN=ON

refs:

  - https://reproducible-builds.org/docs/deterministic-build-systems/
  - https://cmake.org/cmake/help/latest/prop_tgt/BUILD_RPATH.html

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>