csp_service_handler: Disable build timestamp for reproducible builds
While timestamps are useful for tracking when binaries are built, they
introduce non-determinism to the builds. To facilitate reproducible
builds, the build timestamp functionality must be disabled.

Note that this change is API-breaking, and the default behavior is set
to off. Users seeking reproducible builds should explicitly set
CSP_REPRODUCIBLE_BUILDS to 1.

To enable reproducible builds, use the following commands:

waf:
    ./waf configure --enable-reproducible-builds

meson:

    meson setup builddir . -Denable_reproducible_builds=true

cmake:

    make -S . -B builddir -DCSP_REPRODUCIBLE_BUILDS=1

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>