interface: udp: Check return values from pthread functions
This commit addresses the need to check the return values from
pthread_create() and pthread_attr_destroy(). When these functions
return non-zero values, log the corresponding error.

The function csp_if_udp_init() is designed for Linux, and while we
currently assume it doesn't fail (although it can, and proper error
handling should be implemented in a separate pull request), this
commit focuses on logging failures.

Signed-off-by: pr0me <g33sus@gmail.com>
Reviewed-by: Yasushi SHOJI <yashi@spacecubics.com>