interface: can: Fix off-by-one error in csp_can2_rx overflow check
The `csp_packet_t` struct contains two buffers, `header[]` and
`data[]`, consecutively forming a larger buffer. The `header` holds
the CSP header, and the start address of the header is pointed to by
`packet->frame_begin`. To prevent overflow when appending `data` to
the buffer, we need to ensure it doesn't exceed the end.

  header  packet->frame_begin          (end)
  |       |      packet->data          packet->data[sizeof(packet->data)]
  |       |      |                     |
  v       v      v                     v
  +--------------+---------------------+
  |        header|data                 |
  +--------------+---------------------+
          ^                    ^
          |                    | (curr)
          |                    packet->frame_begin[frame_length]
          |
          |<-- frame_length -->|

The length of `data` is `dlc`, and we must ensure that when writing
`data` to the buffer, it won't overflow. That is,

   curr + data < end

Because the end of the buffer is already out-of-band, the error
condition should only be triggered when the total length (`frame_begin
+ frame_length` + `dlc`) is equal to or greater than the end.

This commit fixes the off-by-one error in the overflow check by
adjusting the condition appropriately and reclaiming the unused
location at the very end of the array.

Signed-off-by: pr0me <g33sus@gmail.com>
Reviewed-by: Yasushi SHOJI <yashi@spacecubics.com>